+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Head
 
+- Added: `unspecified: "bottomAlphabetical"` option for `rule-properties-order`.
 - Fixed: `font-weight-notation` does not throw false warnings when `normal` is used in certain ways.
 
 # 4.4.0